refactor: Merge `common` and `metadata` package as a single `script` package
- Metadata package was dealing about the extraction of metadata from
script
- Common package only contained functions for the language of the script
- They both deals with script and make more sense as a single script
package
